enable Travis CI (#5)
The Julia ecosystem has pretty much settled on Travis for CI testing. Travis is free for open source and has pretty nice (community-maintained) Julia support, so it feels like a natural choice for our CI here.

This commit simply adds a .travis.yml file and enables Travis for this repository. In particular the sections of the .travis.yml pertaining to code coverage and docs are commented out until a future PR to add those features.

Build times are currently under a minute, of which actually testing GeohashHilbert takes just a few seconds.